Ajout des pop up pour Inscription et Connexion afin que ca reste dans la page d'accueil
